pull out EGP rest auth calls
